refactor: move things around and deduplicate
